move a bunch of code into service routing table
